+++ otterdog/operations/local_plan_operation.py
@@ -21,19 +21,15 @@ from otterdog.providers.github import Github
 
 
 class LocalPlanOperation(PlanOperation):
-    def __init__(self, suffix: str):
-        super().__init__()
+    def __init__(self, suffix: str, update_webhooks: bool):
+        super().__init__(no_web_ui=False, update_webhooks=update_webhooks)
 
         self.suffix = suffix
         self.other_org: GitHubOrganization | None = None
 
     def pre_execute(self) -> None:
         self.printer.print(f"Printing local diff for configuration at '{self.config.config_file}'")
-        self.printer.print(f"\nActions are indicated with the following symbols:")
-        self.printer.print(f"  {Fore.GREEN}+{Style.RESET_ALL} create")
-        self.printer.print(f"  {Fore.YELLOW}~{Style.RESET_ALL} modify")
-        self.printer.print(f"  {Fore.RED}-{Style.RESET_ALL} extra (missing in definition but available "
-                           f"in other config)")
+        self.print_legend()
 
     def execute(self, org_config: OrganizationConfig) -> int:
         github_id = org_config.github_id